+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Fixed
-- Member count is only displayed for some groups [#2250](https://github.com/rokwire/illinois-app/issues/2250)
+- Member count is only displayed for some groups [#2250](https://github.com/rokwire/illinois-app/issues/2250).
+### Added
+- "Due Date Catalog" to Academics panel drop-down [#2260](https://github.com/rokwire/illinois-app/issues/2260).
 
 ## [4.1.17] - 2022-08-12
 ### Added