+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Member count is only displayed for some groups [#2250](https://github.com/rokwire/illinois-app/issues/2250).
+- Android: set the proper text for snippet view in mapview markers. Show/hide snippet view based on the zoom level [#2268](https://github.com/rokwire/illinois-app/issues/2268).
 ### Added
 - "Due Date Catalog" to Academics panel drop-down [#2260](https://github.com/rokwire/illinois-app/issues/2260).
 ### Changed